typo header not headers

+++ mern/frontend/src/components/WorkoutDetails.jsx
@@ -1,20 +1,33 @@
 import React from 'react'
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 // date-fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 function WorkoutDetails({ workout }) {
+  const { user } = useAuthContext()
   const { dispatch } = useWorkoutsContext()
+
   const handleClick = async () => {
+    if (!user) {
+      return
+    }
+
+    console.log('DELETE', user.token)
     const response = await fetch('/api/workouts/' + workout._id, {
       method: 'DELETE',
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     })
 
     const json = await response.json()
 
     if (response.ok) {
       dispatch({ type: 'DELETE_WORKOUT', payload: json })
+    } else {
+      console.log(response.statusText)
     }
 
     console.log(workout._id)
